lib/parser: rename variable
Rename HaveMarker to Boundary. The HaveMarker doesn't sound properly.
Because the pattern is more like a divider line.

Rename HaveMarker.have() to Boundary.is_inside(). The have function is
really confusing. I think inside a boundary is more proper.

Store pattern inside the Boundary struct. As the variable name has
changed to Boundary, I think the pattern should be part of the bound.

Signed-off-by: Avimitin <avimitin@gmail.com>